\docType{methods}
\name{as.Customer}
\alias{as.Customer}
\alias{as.Customer,data.frame-method}
\alias{as.Customer,list-method}
\title{Convert Objects to Customers}
\usage{
\S4method{as.Customer}{list}(x, ...)

\S4method{as.Customer}{data.frame}(x, ...)
}
\arguments{
  \item{x}{an Object to convert.}

  \item{...}{optional parameters}
}
\description{
Convert Objects to Customers

Convert Objects to Customers
}

