\docType{methods}
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,Customer-method}
\alias{as.data.frame,Node-method}
\alias{as.data.frame,Warehouse-method}
\title{Convert Objects to data.frames}
\usage{
\S4method{as.data.frame}{Node}(x, ...)

\S4method{as.data.frame}{Customer}(x, ...)

\S4method{as.data.frame}{Warehouse}(x, ...)
}
\arguments{
  \item{x}{the Node}

  \item{...}{additional arguments to be passed to or from
  methods.}
}
\description{
Convert Objects to data.frames

Convert Objects to data.frames

Convert Objects to data.frames
}

