\docType{methods}
\name{as.list}
\alias{as.list}
\alias{as.list,Customer-method}
\alias{as.list,Node-method}
\alias{as.list,Warehouse-method}
\title{Convert Objects to lists}
\usage{
\S4method{as.list}{Node}(x, ...)

\S4method{as.list}{Customer}(x, ...)

\S4method{as.list}{Warehouse}(x, ...)
}
\arguments{
  \item{x}{the Node}

  \item{...}{objects, possibly named.}
}
\description{
Convert Objects to lists

Convert Objects to lists

Convert Objects to lists
}

