\docType{methods}
\name{as.Node}
\alias{as.Node}
\alias{as.Node,data.frame-method}
\alias{as.Node,list-method}
\title{Convert Objects to Nodes}
\usage{
\S4method{as.Node}{list}(x, ...)

\S4method{as.Node}{data.frame}(x, ...)
}
\arguments{
  \item{x}{an Object to convert.}

  \item{...}{optional parameters}
}
\description{
Convert Objects to Nodes

Convert Objects to Nodes
}

