\docType{methods}
\name{as.Warehouse}
\alias{as.Warehouse}
\alias{as.Warehouse,data.frame-method}
\alias{as.Warehouse,list-method}
\title{Convert Objects to Warehouses}
\usage{
\S4method{as.Warehouse}{list}(x, ...)

\S4method{as.Warehouse}{data.frame}(x, ...)
}
\arguments{
  \item{x}{an Object to convert.}

  \item{...}{optional parameters}
}
\description{
Convert Objects to Warehouses

Convert Objects to Warehouses
}

