\docType{methods}
\name{calc.polar}
\alias{calc.polar}
\title{Calculate angle of a link connecting two \code{\link{Node}s}}
\usage{
calc.polar(n1, n2, ...)
}
\arguments{
  \item{n1}{an object of class \code{"Node"} or a class
  derived from \code{"Node"}.}

  \item{n2}{an object of class \code{"Node"} or a class
  derived from \code{"Node"}.}

  \item{...}{\emph{Optional Parameters} See Below.}
}
\description{
Calculates the angle of a link connecting two
\code{\link{Node}s}.
}
\section{Optional Parameters (\code{...})}{
  \subsection{used by \code{\link{calc.polar}}}{ \describe{
  \item{...}{No parameters are taken directly.} } }
  \subsection{Forwarded to the follwowing functions}{ You
  may want to check these functions for any other optional
  parameters.  \itemize{ \item{\code{\link{getPhi}}} } }
}
\author{
Dipl. Kfm. Felix Lindemann
\email{felix.lindemann@hs-neu-ulm.de}

Wissenschaftlicher Mitarbeiter Kompetenzzentrum Logistik
Buro ZWEI, 17

Hochschule fur angewandte Wissenschaften Fachhochschule
Neu-Ulm | Neu-Ulm University Wileystr. 1

D-89231 Neu-Ulm

Phone +49(0)731-9762-1437 Web
\url{www.hs-neu-ulm.de/felix-lindemann/}
\url{http://felixlindemann.blogspot.de}
}

