\docType{methods}
\name{drawLinks}
\alias{drawLinks}
\alias{drawLinks,GeoSituation-method}
\title{drawLinks}
\usage{
drawLinks(object, ...)

\S4method{drawLinks}{GeoSituation}(object, ...)
}
\arguments{
  \item{object}{the \code{\link{GeoSituation}-Object(s)} to
  be plotted}

  \item{...}{list of optional values (see below)}
}
\description{
drawLinks
}
\details{
Optional Parameters for this Methdod.
}
\section{optional arguments}{
  \describe{ \item{\code{withlabels}:}{
  \code{\link{logical}} default = \code{TRUE}. If
  \code{TRUE}, labels will be printed for each Node.  }
  \item{\code{pch}:}{ \code{\link{numeric}} default = 21.
  Defines the point-Type to be used.  }
  \item{\code{point.cex}:}{ \code{\link{numeric}} default =
  3. Defines the size of the points.  }
  \item{\code{font.cex}:}{ \code{\link{numeric}} default =
  3. Defines the size of the font.  } \item{\code{bg}:}{
  \code{\link{numeric} or \link{character}} default =
  \code{"white"}. The background-color the node-background
  should be printed in.  } \item{\code{col}:}{
  \code{\link{numeric} or \link{character}} default =
  \code{"black"}. The foreground-color the node-outline
  should be printed in.  } }
}

