\docType{methods}
\name{drawTPP}
\alias{drawTPP}
\alias{drawTPP,GeoSituation-method}
\title{drawTPP}
\usage{
drawTPP(object, ...)

\S4method{drawTPP}{GeoSituation}(object, ...)
}
\arguments{
  \item{object}{the \code{\link{GeoSituation}-Object(s)} to
  be plotted}

  \item{...}{list of optional values (see below)}
}
\description{
drawTPP
}
\details{
Optional Parameters for this Methdod.
}
\section{optional arguments}{
  \describe{ \item{\code{withlabels}:}{
  \code{\link{logical}} default = \code{TRUE}. If
  \code{TRUE}, labels will be printed for each Node.  } }
}

