\docType{methods}
\name{drawWLP}
\alias{drawWLP}
\alias{drawWLP,GeoSituation-method}
\title{drawWLP}
\usage{
drawWLP(object, ...)

\S4method{drawWLP}{GeoSituation}(object, ...)
}
\arguments{
  \item{object}{the \code{\link{GeoSituation}-Object(s)} to
  be plotted}

  \item{...}{list of optional values (see below)}
}
\description{
drawWLP
}
\details{
Optional Parameters for this Methdod.
}
\section{optional arguments}{
  \describe{ \item{\code{...}:}{ Is forwarded to
  \code{\link{drawTPP}}.  } }
}

