\docType{methods}
\name{$}
\alias{$}
\alias{$,GeoSituation-method}
\alias{$,Link-method}
\alias{$,Node-method}
\title{Extract Methods}
\usage{
\S4method{$}{Node}(x, name)

\S4method{$}{Link}(x, name)

\S4method{$}{GeoSituation}(x, name)
}
\arguments{
  \item{name}{Name of Attribute.}

  \item{x}{ object from which to extract element(s) or in
  which to replace element(s).  }
}
\description{
Extract Methods

Extract Method

Extract Methods
}

