\docType{methods}
\name{[}
\alias{[}
\alias{[,Customer-method}
\alias{[,GeoSituation-method}
\alias{[,Node-method}
\alias{[,Warehouse-method}
\title{Extract Methods}
\usage{
\S4method{[}{Node}(x, i, j, drop)

\S4method{[}{Customer}(x, i, j, drop)

\S4method{[}{Warehouse}(x, i, j, drop)

\S4method{[}{GeoSituation}(x, i, j, drop)
}
\arguments{
  \item{i}{Row Index}

  \item{j}{Name or Column index}

  \item{drop}{Optional value for Drop-Levels.}

  \item{x}{ object from which to extract element(s) or in
  which to replace element(s).  }
}
\description{
Extract Methods

Extract Methods

Extract Methods

Extract Methods
}

