\docType{methods}
\name{$}
\alias{$}
\alias{$,Customer-method}
\alias{$,GeoSituation-method}
\alias{$,Link-method}
\alias{$,Node-method}
\alias{$,Warehouse-method}
\title{Extract Method}
\usage{
\S4method{$}{Customer}(x, name)

\S4method{$}{GeoSituation}(x, name)

\S4method{$}{Link}(x, name)

\S4method{$}{Node}(x, name)

\S4method{$}{Warehouse}(x, name)
}
\arguments{
  \item{x}{ object from which to extract element(s) or in
  which to replace element(s).  }

  \item{name}{ A literal character string or a \link{name}
  (possibly \link{backtick} quoted).  For extraction, this
  is normally (see under \sQuote{Environments}) partially
  matched to the \code{\link{names}} of the object.  }
}
\description{
Extract Method

Extract Method

Extract Method

Extract Method
}

