\docType{methods}
\name{is.Customer}
\alias{is.Customer}
\alias{is.Customer,Customer-method}
\title{Checks if an Object is as Customer}
\usage{
is.Customer(x, ...)

\S4method{is.Customer}{Customer}(x, ...)
}
\arguments{
  \item{x}{the object to be checked for class
  \code{\link{Customer}}}

  \item{...}{optional Parameters (not supported)}
}
\description{
Checks if an Object is as Customer
}

