\docType{methods}
\name{is.Node}
\alias{is.Node}
\alias{is.Node,Node-method}
\title{Checks if an Object is as Node}
\usage{
is.Node(x, ...)

\S4method{is.Node}{Node}(x, ...)
}
\arguments{
  \item{x}{the Node}

  \item{...}{optional Parameters (not supported)}
}
\description{
Checks if an Object is as Node
}

