\docType{methods}
\name{is.Warehouse}
\alias{is.Warehouse}
\alias{is.Warehouse,Warehouse-method}
\title{Checks if an Object is as Warehouse}
\usage{
is.Warehouse(x, ...)

\S4method{is.Warehouse}{Warehouse}(x, ...)
}
\arguments{
  \item{x}{the Warehouse}

  \item{...}{optional Parameters (not supported)}
}
\description{
Checks if an Object is as Warehouse
}

