\docType{methods}
\name{length}
\alias{length}
\alias{length,Customer-method}
\alias{length,Node-method}
\alias{length,Warehouse-method}
\title{How many Nodes are included?}
\usage{
\S4method{length}{Node}(x)

\S4method{length}{Customer}(x)

\S4method{length}{Warehouse}(x)
}
\arguments{
  \item{x}{the Nodes}
}
\description{
How many Nodes are included?

How many Customers are included?

How many Warehouses are included?
}

