\docType{methods}
\name{$<-}
\alias{$<-}
\alias{$<-,GeoSituation-method}
\alias{$<-,Link-method}
\alias{$<-,Node-method}
\title{Set Methods}
\usage{
\S4method{$}{Node}(x, name) <- value

\S4method{$}{Link}(x, name) <- value

\S4method{$}{GeoSituation}(x, name) <- value
}
\arguments{
  \item{name}{Name of parameter to change}

  \item{value}{new Value.}

  \item{x}{ object from which to extract element(s) or in
  which to replace element(s).  }
}
\description{
Set Methods

Set Method

Set Methods
}

