\name{[<-}
\alias{[<-}
\alias{[<-,Customer-method}
\alias{[<-,Node-method}
\alias{[<-,Warehouse-method}
\title{Set Methods}
\arguments{
  \item{x}{The Node}

  \item{i}{row-Index}

  \item{j}{Name or Column index}

  \item{x}{The Customer}

  \item{i}{row-Index}

  \item{j}{Name or Column index}

  \item{i}{row-Index}

  \item{j}{Name or Column index}
}
\description{
Set Methods

Set Methods

Set Methods
}

