\docType{methods}
\name{show}
\alias{show}
\alias{show,Customer-method}
\alias{show,Link-method}
\alias{show,Node-method}
\alias{show,Warehouse-method}
\title{show Method}
\usage{
\S4method{show}{Node}(object)

\S4method{show}{Customer}(object)

\S4method{show}{Warehouse}(object)

\S4method{show}{Link}(object)
}
\arguments{
  \item{object}{Object can be of type \code{\link{Node}}}

  \item{object}{Object can be of type
  \code{\link{Customer}}}

  \item{object}{Object can be of type
  \code{\link{Warehouse}}}

  \item{object}{Object can be of type \code{\link{Link}}}
}
\description{
Object can be of type \code{\link{Node}}

Object can be of type \code{\link{Customer}}

Object can be of type \code{\link{Warehouse}}

Object can be of type \code{\link{Link}}
}

