\docType{methods}
\name{plotGeoSituation}
\alias{plotGeoSituation}
\alias{plotGeoSituation,GeoSituation-method}
\title{Plot an Object of Type \code{\link{GeoSituation}}}
\usage{
plotGeoSituation(object, ...)

\S4method{plotGeoSituation}{GeoSituation}(object, ...)
}
\arguments{
  \item{object}{an object of class
  \code{\link{GeoSituation}}}

  \item{...}{Additional argument list that might not ever
  be used.}
}
\value{
The updated object of class \code{\link{GeoSituation}}.
}
\description{
This method should simplify the use of HNUORTools.
}
\note{
for citing use: Felix Lindemann (2014). HNUORTools:
Operations Research Tools. R package version 1.1-0.
\url{http://felixlindemann.github.io/HNUORTools/}.
}
\author{
Dipl. Kfm. Felix Lindemann
\email{felix.lindemann@hs-neu-ulm.de}

Wissenschaftlicher Mitarbeiter Kompetenzzentrum Logistik
Buro ZWEI, 17

Hochschule fur angewandte Wissenschaften Fachhochschule
Neu-Ulm | Neu-Ulm University Wileystr. 1

D-89231 Neu-Ulm

Phone +49(0)731-9762-1437 Web
\url{www.hs-neu-ulm.de/felix-lindemann/}
\url{http://felixlindemann.blogspot.de}
}

